root/requirements.txt
asgiref==3.7.2
bleach==6.0.0
Django==4.2.5
django-classy-tags==4.1.0
django-extensions==3.2.3
django-js-asset==2.1.0
django-mptt==0.14.0
django-nyt==1.3
django-sekizai==4.1.0
django-taggit==5.0.1
django-taggit-helpers==0.1.4
Markdown==3.3.7
Pillow==10.0.1
psycopg2-binary==2.9.9
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
tinycss2==1.1.1
typing_extensions==4.8.0
Unidecode==1.3.7
uWSGI==2.0.22
webencodings==0.5.1
wiki==0.10
